Fix input parameter type for `xnn_set1_or_load_s32`.
PiperOrigin-RevId: 652461844

+++ src/xnnpack/simd/s32-neon.h
@@ -49,7 +49,7 @@ static XNN_INLINE xnn_simd_s32_t xnn_set1_s32(int32_t v) {
   return vld1q_dup_s32(&v);
 }
 
-static XNN_INLINE xnn_simd_s32_t xnn_set1_or_load_s32(const float* v) {
+static XNN_INLINE xnn_simd_s32_t xnn_set1_or_load_s32(const int32_t* v) {
   return vld1q_dup_s32(v);
 }
